Splitting up utilities to a package
